requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.13.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.13.0
frozendict==2.4.6
frozenlist==1.8.0
idna==3.10
multidict==6.7.0
multitasking==0.0.12
nulltype==2.3.1
numpy==2.3.2
pandas==2.3.2
peewee==3.18.2
plaid-python==37.1.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==6.32.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.8
twilio==9.8.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.65
